--- a/cellxgene_schema_cli/requirements.txt
+++ b/cellxgene_schema_cli/requirements.txt
@@ -2,7 +2,7 @@ anndata==0.11.2
 cellxgene-ontology-guide==1.3.0 # update before a schema migration
 click<9
 Cython<4
-dask==2024.12.0
+dask[array]==2024.12.0
 numpy<3
 pandas>2,<3
 PyYAML<7