@nayib-jose-gloria 2025-01-15
> I did not do the bench mark against the custom dask config since both `threads_per_worker`, `distributed.worker.memory.limit` would be unused parameters based off the source code. Testing with `num_works==1` would essentially be testing against the custom Dask config, but to be thorough, i'll run it with the original config.

oh I think I misunderstood part of the description; I thought the custom config was adding dask overhead that was hindering performance; but I see now that you meant dev overhead to maintain the config. You can skip further benchmarking